Semantic and rename changes to introduction section
Some tags in the introduction section had been changed to improve the semantic meaning of the page in this section. Also the names of some classes were changed due to its confusing name meaning.

Other changes:
tool-section-v1.css file was deleted. This file is from an older version of the page were i wasn't using git for version control. Due to the recent migration to a git repository, it isn't necessary anymore.