Semantic and rename changes to tool section
Some tags in the tool section had been replaced to improve the semantic meaning of the page. Also this changes bring some visual bugs with the recently implemented animations. The following changes were implemented to fix these:
-Now the entry animation of a panel feature is applied in the <li> tag, instead of the <a> tag

Another changes:
The delay of the entry animations of the panel features is now lower (from 5% for each one to 4%) so these now entry faster