Animations changed in tool section
Scroll animations have been replaced in the tool section.
To improve the performance and avoid possible future errors, the direction of the animations on the feature list elements has been changed (from: bottom to top, to: right to left)
and the animation distance has been decreased as well.
As a design improvement of this new design change, the fade-in key-frame has been implemented, so the elements not only slide but also have a fade-in effect on entry.