Element animation doesn't restart after it was performed once
Now the entry animations doesn't restart when it was performed once, when the element enters in the view-port an performs its animation, it stays in its final position even if it gets out of view-port again.
This due to a slower navigation experience every time someone wants to return to a section, causing as consequence, an awful user navigation experience.